% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_sp_list}
\alias{data_sp_list}
\title{Import WWF Wildfinder species data}
\usage{
data_sp_list()
}
\value{
A tibble containing species IDs and taxonomic info
}
\description{
Import WWF Wildfinder species data
}
